Rename durconn -> stanmsg
